requirements.txt
torch
torchvision
torchaudio
--index-url https://download.pytorch.org/whl/cu118
torchmetrics
seaborn
nltk
Unidecode
notebook
matplotlib
pytorch-lightning
librosa
scipy
inflect
pytest
ipywidgets
conformer
phonemizer
triton==2.0.0.dev20221105 --install-option="--no-deps"
transformers
xformers
